pyproject.toml
[project]
name = "proyecto1"
version = "0.1.0"
description = "Simulador de Maquina de Turing - Sucesion de Fibonacci"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pyyaml>=6.0.3",
]
